requirements.txt
aider-chat==0.86.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
build==1.4.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
chromadb==1.4.1
click==8.2.1
coloredlogs==15.0.1
ConfigArgParse==1.7.1
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
dataclasses-json==0.6.7
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
docx2txt==0.9
durationpy==0.10
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.12.19
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.4
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.3.1
grep-ast==0.9.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==35.0.0
langchain==1.2.7
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-pinecone==0.2.13
langchain-tavily==0.2.17
langchain-text-splitters==1.1.0
langchain-upstage==0.7.6
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
litellm==1.75.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
mmh3==5.2.0
mpmath==1.3.0
mslex==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.16.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
oslex==0.1.3
overrides==7.7.0
packaging==24.2
pathspec==0.12.1
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.7
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.6.2
pypdfium2==5.3.0
pyperclip==1.9.0
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
scipy==1.15.3
setuptools==80.10.2
shellingham==1.5.4
shtab==1.7.2
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.46
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.10.0
tokenizers==0.20.3
torch==2.10.0
tqdm==4.67.1
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.9.0
tree-sitter-yaml==0.7.1
triton==3.6.0
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
